refactor: divide TrainCars from Player
