+++ src/main/scala/controller/ViewController.scala
@@ -47,7 +47,7 @@ private object ViewController:
             case _ => throw new IllegalStateException("Unhandled mechanic")
         )
       )
-      gameView.updatePlayerInfo(currentPlayer.id, currentPlayer.trains.trainCars)
+      gameView.updatePlayerInfo(currentPlayer.id, currentPlayer.trains)
       gameView.addHandView(handView)
       gameView.updateObjective(currentPlayerObjective)
       gameView.initPlayerScores(currentPlayerScores)
@@ -69,7 +69,7 @@ private object ViewController:
         case END_GAME => gameView.endGame(currentPlayerScores)
         case _ => ()
       handView.updateHand(currentHandCardsView)
-      gameView.updatePlayerInfo(currentPlayer.id, currentPlayer.trains.trainCars)
+      gameView.updatePlayerInfo(currentPlayer.id, currentPlayer.trains)
       gameView.updateHandView(handView)
       gameView.updateCompletionCheckBox(currentPlayer.objective.completed)
       gameView.updateObjective(currentPlayerObjective)
